OPENDJ-1802 Make ByteStringBuilder methods more intentional
ByteStringBuilder.java:
Changed:
- append(byte) to appendByte(int)
- append(short) to appendShort(int)
- append(int) to appendInt(int)
- append(long) to appendLong(long)
- append(Object) to appendObject(Object)
- append(String) to appendUtf8(String)
- append(char[]) to appendUtf8(char[])
- All remaining append() to appendBytes().
I changed appendByte() and appendShort() to accept int parameters.
This avoids continuously casting, at the expense of a possible loss of the high bits.

ByteStringBuilder.java: appendBytes(), appendUtf8(), appendObject()

+++ opendj-server-legacy/src/main/java/org/opends/server/replication/protocol/LDAPUpdateMsg.java
@@ -221,7 +221,7 @@ public abstract class LDAPUpdateMsg extends UpdateMsg
      */
     final ByteArrayBuilder builder = new ByteArrayBuilder();
     builder.appendByte(msgType);
-    builder.appendByte((byte) protocolVersion);
+    builder.appendByte(protocolVersion);
     builder.appendCSNUTF8(csn);
     builder.appendDN(dn);
     builder.appendString(entryUUID);
